requirements.txt
requests==2.31.0
pytest==8.3.3
allure-pytest==2.13.5
